fix: checking if the old save file exists
when we open the game in a different scene in the first time (when we don't start the game in the Initialization scene), the error is raised.

+++ UOP1_Project/Assets/Scripts/SaveSystem/FileManager.cs
@@ -48,6 +48,12 @@ public static class FileManager
 			{
 				File.Delete(newFullPath);
 			}
+
+			if (!File.Exists(fullPath))
+			{
+				return false;
+			}
+			
 			File.Move(fullPath, newFullPath);
 		}
 		catch (Exception e)
